project/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
lorem==0.1.1
names==0.3.0
numpy==1.21.4
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.2.4
